+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- Remove method AckType
 - Add separate types PubComp, PubRec, PubRel
 - Returned *Malformed from Publish.WellFormed
 - Remove types Handler, PubHandler and Middleware